requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-gis==1.0
psycopg==3.1.19
psycopg-binary==3.1.19
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.2
